Question - 2 [ Set 3 ] Description
Q2 - Write a program to demonstrate fork where parent and child run same codes and parent process should be executed first.

Fork system call is used for creating a new process, which is called child process, which runs concurrently with the process that makes the fork() call (parent process).  From the question we have to demonstrate the fork program where parent and child process run same codes and parent process should be executed first. 

The codes is written above where both parent and child run together and the final output is " Parent Complete "